+++ python/lsst/ts/cRIOpy/VMS/PSDWidget.py
@@ -19,21 +19,15 @@
 
 __all__ = ["PSDWidget"]
 
-from ..GUI.TimeChart import AbstractChart
-from .ChartView import ChartView
-from .Unit import units, coefficients
-from ..GUI.CustomLabels import DockWindow
+from .CacheWidget import CacheWidget
 
-from PySide2.QtCore import Qt, Slot, QPointF
-from PySide2.QtCharts import QtCharts
+from PySide2.QtCore import Qt, QPointF
 
-import concurrent.futures
 import numpy as np
 import time
-from lsst.ts.utils import make_done_future
 
 
-class PSDWidget(DockWindow):
+class PSDWidget(CacheWidget):
     """Display signal PSD.
 
     Parameters
@@ -49,90 +43,10 @@ class PSDWidget(DockWindow):
     """
 
     def __init__(self, title, cache, SAMPLE_TIME, toolBar, channels=[]):
-        super().__init__(title)
-        self.setObjectName(title)
-        self.SAMPLE_TIME = SAMPLE_TIME
-        self.toolBar = toolBar
+        super().__init__(title, cache, SAMPLE_TIME, toolBar, channels)
 
-        self.chart = AbstractChart()
-
-        self.updateTask = make_done_future()
-
-        self.update_after = 0
-
-        self.cache = cache
-
-        self.chartView = ChartView(self.chart, QtCharts.QLineSeries)
-        self.chartView.updateMaxSensor(self.cache.sensors())
-        self.chartView.axisChanged.connect(self.axisChanged)
-        self.chartView.unitChanged.connect(self.unitChanged)
-        for channel in channels:
-            self.chartView.addSerie(str(channel[0]) + " " + channel[1])
-
-        self.coefficient = 1
-        self.unit = units[0]
-        self.setupAxes = True
-        self._setupAxes()
-
-        self.setWidget(self.chartView)
-
-    @Slot(bool, bool)
-    def axisChanged(self, logX, logY):
-        self.setupAxes = True
-
-    @Slot(str)
-    def unitChanged(self, unit):
-        self.coefficient = coefficients(unit)
-        self.unit = unit
-        self.setupAxes = True
-
-    def _setupAxes(self):
-        for a in self.chart.axes():
-            self.chart.removeAxis(a)
-
-        if len(self.chart.series()) == 0:
-            return
-
-        if self.chartView.logX:
-            xAxis = QtCharts.QLogValueAxis()
-        else:
-            xAxis = QtCharts.QValueAxis()
-
-        if self.chartView.logY:
-            yAxis = QtCharts.QLogValueAxis()
-        else:
-            yAxis = QtCharts.QValueAxis()
-
-        xAxis.setTitleText("Frequency (Hz)")
-        yAxis.setTitleText("PSD ((" + self.unit + ")<sup>2</sup> Hz <sup>-1</sup>)")
-
-        self.chart.addAxis(xAxis, Qt.AlignBottom)
-        self.chart.addAxis(yAxis, Qt.AlignLeft)
-
-        for s in self.chart.series():
-            s.attachAxis(xAxis)
-            s.attachAxis(yAxis)
-
-        self.chart.axes(Qt.Horizontal)[0].setGridLineVisible(True)
-        self.chart.axes(Qt.Horizontal)[0].setMinorGridLineVisible(True)
-
-        self.chart.legend().setAlignment(Qt.AlignTop)
-
-        self.frequencyChanged(*self.toolBar.getFrequencyRange())
-
-        self.setupAxes = False
-
-    @Slot(int, int, float, float)
-    def cacheUpdated(self, index, length, startTime, endTime):
-        """Process and plot data.
-
-        Parameters
-        ----------
-        index : `int`
-        length : `int`
-        startTime : `float`
-        endTime : `float`
-        """
+    def plotAll(self):
+        """Plot all signals. Run as task in a thread."""
 
         def downsample(psd, N):
             """Downsample PSD so no too many points are plot. Replace PSD with
@@ -213,42 +127,16 @@ class PSDWidget(DockWindow):
 
             return min(psd), max(psd)
 
-        def plotAll():
-            """Plot all signals. Run as task in a thread."""
-            #   min_psd, max_psd = plot(
-            #   0,
-            #   np.mean(
-            #      [self.cache[s + self.samples[0]] for s in ["1", "2", "3"]],
-            #      axis=0,
-            #   ),
-            #   )
-            #   self.chart.axes(Qt.Vertical)[0].setRange(min_psd, max_psd)
-            min_psd = []
-            max_psd = []
-            for s in self.chart.series():
-                min_p, max_p = plot(s, self.cache[s.name()])
-                min_psd.append(min_p)
-                max_psd.append(max_p)
+        min_psd = []
+        max_psd = []
+        for s in self.chart.series():
+            min_p, max_p = plot(s, self.cache[s.name()])
+            min_psd.append(min_p)
+            max_psd.append(max_p)
 
-            if len(min_psd) > 0:
-                if len(self.chart.axes(Qt.Vertical)) == 0:
-                    self.setupAxes = True
-                else:
-                    self.chart.axes(Qt.Vertical)[0].setRange(min(min_psd), max(max_psd))
-            self.update_after = time.monotonic() + 0.5
-
-        if self.setupAxes is True:
-            self._setupAxes()
-            self.update_after = 0
-
-        if self.update_after < time.monotonic() and self.updateTask.done():
-            with concurrent.futures.ThreadPoolExecutor() as pool:
-                self.updateTask = pool.submit(plotAll)
-
-    @Slot(float, float)
-    def frequencyChanged(self, low, high):
-        if len(self.chart.series()) == 0:
-            self.setupAxes = True
-            return
-
-        self.chart.axes(Qt.Horizontal)[0].setRange(low, high)
+        if len(min_psd) > 0:
+            if len(self.chart.axes(Qt.Vertical)) == 0:
+                self.setupAxes = True
+            else:
+                self.chart.axes(Qt.Vertical)[0].setRange(min(min_psd), max(max_psd))
+        self.update_after = time.monotonic() + 0.5
